Don't try to save the patched item if the target is a directory and the patch only applied property changes to prevent an error dialog from popping up. (Fixes issue #279) : applying patch with property changes shows error
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@22831 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseMerge/MainFrm.cpp
@@ -1148,23 +1148,26 @@ void CMainFrame::PatchSave()
     {
         m_Patch.PatchPath(m_Data.m_mergedFile.GetFilename());
     }
-    int saveret = SaveFile(m_Data.m_mergedFile.GetFilename());
-    if (saveret==0)
+    if (!PathIsDirectory(m_Data.m_mergedFile.GetFilename()))
     {
-        // file was saved with 0 lines, remove it.
-        m_Patch.RemoveFile(m_Data.m_mergedFile.GetFilename());
-        // just in case
-        DeleteFile(m_Data.m_mergedFile.GetFilename());
-    }
-    m_Data.m_mergedFile.StoreFileAttributes();
-    if (m_Data.m_mergedFile.GetFilename() == m_Data.m_yourFile.GetFilename())
-        m_Data.m_yourFile.StoreFileAttributes();
-    if ((bDoesNotExist)&&(DWORD(CRegDWORD(_T("Software\\TortoiseMerge\\AutoAdd"), TRUE))))
-    {
-        // call TortoiseProc to add the new file to version control
-        CString cmd = _T("/command:add /noui /path:\"");
-        cmd += m_Data.m_mergedFile.GetFilename() + _T("\"");
-        CAppUtils::RunTortoiseProc(cmd);
+        int saveret = SaveFile(m_Data.m_mergedFile.GetFilename());
+        if (saveret==0)
+        {
+            // file was saved with 0 lines, remove it.
+            m_Patch.RemoveFile(m_Data.m_mergedFile.GetFilename());
+            // just in case
+            DeleteFile(m_Data.m_mergedFile.GetFilename());
+        }
+        m_Data.m_mergedFile.StoreFileAttributes();
+        if (m_Data.m_mergedFile.GetFilename() == m_Data.m_yourFile.GetFilename())
+            m_Data.m_yourFile.StoreFileAttributes();
+        if ((bDoesNotExist)&&(DWORD(CRegDWORD(_T("Software\\TortoiseMerge\\AutoAdd"), TRUE))))
+        {
+            // call TortoiseProc to add the new file to version control
+            CString cmd = _T("/command:add /noui /path:\"");
+            cmd += m_Data.m_mergedFile.GetFilename() + _T("\"");
+            CAppUtils::RunTortoiseProc(cmd);
+        }
     }
 }
